+++ AetheryteLinkInChat/Solver/AetheryteSolver.cs
@@ -9,7 +9,7 @@ using Divination.AetheryteLinkInChat.Config;
 using Lumina.Excel;
 using Lumina.Excel.GeneratedSheets;
 
-namespace Divination.AetheryteLinkInChat;
+namespace Divination.AetheryteLinkInChat.Solver;
 
 public class AetheryteSolver(IDataManager dataManager)
 {
@@ -39,7 +39,7 @@ public class AetheryteSolver(IDataManager dataManager)
                     DalamudLog.Log.Verbose("P1 = ({X1}, {Y1}), P2 = ({X2}, {Y2})", x, y, markerX, markerY);
 
                     DalamudLog.Log.Verbose("path = {S}", path);
-                    if (path is AetheryteTeleportPath {Aetheryte.AethernetGroup: > 0})
+                    if (path is AetheryteTeleportPath { Aetheryte.AethernetGroup: > 0 })
                     {
                         DalamudLog.Log.Verbose("skip distance calculation: this is aethernet: {S}", path);
                     }
